Update src/javascript/ZeroClipboard.js
The glue function actualy does not consider the element hierarchie in the dom. Its aparently just appending the layer containing the flash to the parent element of the given "glue" element. If this is i.e. a TD or a SPAN in an anchor, it renders invalid HTML which leads to the IE error as soon as it gets assigned (innerHML = xxx)